BB-68 refactor the name of Calculation
