+++ src/test/java/com/butbetter/applicationservices/caluapi/CalculatorServiceTestUsingHeroku.java
@@ -1,6 +1,6 @@
 package com.butbetter.applicationservices.caluapi;
 
-import com.butbetter.applicationservices.caluapi.service.Calculator;
+import com.butbetter.applicationservices.caluapi.service.CalculatorService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,10 +11,10 @@ import org.springframework.web.client.ResourceAccessException;
 import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
-class CalculatorTestUsingHeroku {
+class CalculatorServiceTestUsingHeroku {
 
     @Autowired
-    private Calculator calculator;
+    private CalculatorService calculatorService;
 
     //integration Testing on heroku:
 
@@ -22,26 +22,26 @@ class CalculatorTestUsingHeroku {
     @Test
     @DisplayName("test calc API on heroku with the values 100 10")
     void testCalcAPIReturnValueFor10percentFrom100(){
-        assertEquals(110,calculator.calculateVATofPrice(100,10));
+        assertEquals(110, calculatorService.calculateVATofPrice(100,10));
     }
 
     @Test
     @DisplayName("test calc API with the values -100 10")
     void testCalcAPIReturnForIllegalInputPrice(){
-        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculator.calculateVATofPrice(-100,10));
+        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculatorService.calculateVATofPrice(-100,10));
     }
 
     @Test
     @DisplayName("test calc API with the values 100 -10")
     void testCalcAPIReturnForIllegalInputPercent(){
-        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculator.calculateVATofPrice(100,-10));
+        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculatorService.calculateVATofPrice(100,-10));
     }
 
     //this test asserts to true if heroku calcApi is not running
     @Test
     @DisplayName("test a not running calc API with the values 100 10")
     void testCalcAPIReturnIfCalcAPINotAvailable(){
-        assertThrows(ResourceAccessException.class, ()-> calculator.calculateVATofPrice(100,10));
+        assertThrows(ResourceAccessException.class, ()-> calculatorService.calculateVATofPrice(100,10));
     }
