+++ src/test/java/com/butbetter/applicationservices/caluapi/CalculatorServiceTestUsingLocalhost.java
@@ -1,6 +1,6 @@
 package com.butbetter.applicationservices.caluapi;
 
-import com.butbetter.applicationservices.caluapi.service.Calculator;
+import com.butbetter.applicationservices.caluapi.service.CalculatorService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,10 +11,10 @@ import org.springframework.web.client.ResourceAccessException;
 import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
-class CalculatorTestUsingLocalhost {
+class CalculatorServiceTestUsingLocalhost {
 
     @Autowired
-    private Calculator calculator;
+    private CalculatorService calculatorService;
 
     //integration Testing on localhost:
 
@@ -22,30 +22,30 @@ class CalculatorTestUsingLocalhost {
     @Test
     @DisplayName("test calc API with the values 100 10")
     void testCalcAPIReturnValueFor10percentFrom100(){
-        calculator.setCalcApiUrl("http://localhost:8080");
-        assertEquals(110,calculator.calculateVATofPrice(100,10));
+        calculatorService.setCalcApiUrl("http://localhost:8080");
+        assertEquals(110, calculatorService.calculateVATofPrice(100,10));
     }
 
     @Test
     @DisplayName("test calc API with the values -100 10")
     void testCalcAPIReturnForIllegalInputPrice(){
-        calculator.setCalcApiUrl("http://localhost:8080");
-        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculator.calculateVATofPrice(-100,10));
+        calculatorService.setCalcApiUrl("http://localhost:8080");
+        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculatorService.calculateVATofPrice(-100,10));
     }
 
     @Test
     @DisplayName("test calc API with the values 100 -10")
     void testCalcAPIReturnForIllegalInputPercent(){
-        calculator.setCalcApiUrl("http://localhost:8080");
-        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculator.calculateVATofPrice(100,-10));
+        calculatorService.setCalcApiUrl("http://localhost:8080");
+        assertThrows(HttpClientErrorException.BadRequest.class, ()-> calculatorService.calculateVATofPrice(100,-10));
     }
 
     //this test asserts to true if localhost calcApi not running
     @Test
     @DisplayName("test not running calc API with the values 100 10")
     void testCalcAPIReturnIfCalcAPINotAvailable(){
-        calculator.setCalcApiUrl("http://localhost:8080");
-        assertThrows(ResourceAccessException.class, ()-> calculator.calculateVATofPrice(100,10));
+        calculatorService.setCalcApiUrl("http://localhost:8080");
+        assertThrows(ResourceAccessException.class, ()-> calculatorService.calculateVATofPrice(100,10));
     }
